Added support for `img` tags nested inside `a`. These images inherit the link from their parent and are clickable.
Flux will now run when jQuery is in compatibility mode.

Added basic inbuilt next/prev buttons. Pass in `control: true` to the constructor.

Flux no longer silently fails on unsupported browsers, it builds the slider but just switches from one image to the other.

Refactored setup code so that it works on IE.

Added a lock to ensure that the `finish()` function of `flux.transition` only ever fires once per transition.

Added support for `swipeLeft` & `swipeRight` when used with Zepto.js.

Forced hardware acceleration on the `slide` transition so that it works better on iOS.

The `next()` & `prev()` functions now pass on the `left`/`right` param for supported transitions.